changed budget related properties from int to double

+++ game/properties/Bank.java
@@ -1,3 +1,4 @@
+
 package game.properties;
 
 import game.Player;
@@ -12,7 +13,7 @@ public class Bank extends BankProperties{
      */
     public final static int[] atFields = {21};
     private double INVEST_RATIO = 0.5;
-    private int PROFIT_RATIO = 2;
+    private double PROFIT_RATIO = 2.0;
 
     public Bank(int atField){
         this.atField = atField;
@@ -24,8 +25,8 @@ public class Bank extends BankProperties{
      */
     public void invest(Player player){
         double investedMoney = player.getCash() * INVEST_RATIO;
-        player.setInvestedMoney((int)investedMoney);
-        player.setCash(player.getCash() - (int)investedMoney);
+        player.setInvestedMoney(investedMoney);
+        player.setCash(player.getCash() - investedMoney);
         player.setNetWorth(player.getPropertyWorth() + player.getCash());//update net worth
     }
 
@@ -36,7 +37,7 @@ public class Bank extends BankProperties{
      */
     public void profit(Player player){
         if(player.getInvestedMoney() != 0){
-            int addedCash = player.getInvestedMoney() * PROFIT_RATIO;
+            double addedCash = player.getInvestedMoney() * PROFIT_RATIO;
             player.setCash(player.getCash() + addedCash);
             player.setNetWorth(player.getPropertyWorth() + player.getCash());//update net worth
             player.setInvestedMoney(0);//update invested money
